backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
stripe==7.10.0
pillow==10.1.0
pytest==7.4.3
pytest-cov==4.1.0
email-validator==2.1.0.post1
gunicorn==21.2.0